// frontend/src/api/sdk/api.ts
export interface Token {
    'access_token': string;
    'token_type': string;
}
export interface User {
    'username': string;
    'password': string;
    'disabled'?: boolean | null;
}